refactor(app): extract RawData into app/raw.rs
Pure data bag, moved verbatim and re-exported. Part of #270.
